Extract config tests into dedicated module
